+++ web/src/pages/Dashboard/SubDashboard.tsx
@@ -1,14 +1,8 @@
-import React, { useState } from 'react';
-import {
-    IDashboardMission,
-    DashboardMissionList,
-} from './DashboardMission.type';
+import React from 'react';
 import './Dashboard.scss';
 import DashboardMission from './DashboardMission';
 
 export default function SubDashboard() {
-    const [missionList] = useState(DashboardMissionList as IDashboardMission[]);
-
     return (
         <div className="dashboard-pages">
             <div className="page-info">
@@ -20,7 +14,7 @@ export default function SubDashboard() {
             </div>
 
             <div className="assigned-missions">
-                <DashboardMission list={missionList} />
+                <DashboardMission />
             </div>
         </div>
     );
